test(dataframe): move #12147 regression to separate test file\n\nPlace dedicated regression test in 	est_concat_issue_12147.py for clarity and to keep tests independent.
